// src/helpers/schema.tsx
export type AdminRole = "content_admin" | "event_admin" | "super_admin";
// ...
export type UserRole = "admin" | "member";